move GpuSaveAsHiveFile into separate file
