+++ sql-plugin/src/main/spark311/scala/org/apache/spark/sql/hive/rapids/shims/GpuSaveAsHiveFile.scala
@@ -1,0 +1,75 @@
+/*
+ * Copyright (c) 2023, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*** spark-rapids-shim-json-lines
+{"spark": "311"}
+{"spark": "312"}
+{"spark": "313"}
+{"spark": "320"}
+{"spark": "321"}
+{"spark": "321cdh"}
+{"spark": "321db"}
+{"spark": "322"}
+{"spark": "323"}
+{"spark": "330"}
+{"spark": "330cdh"}
+{"spark": "330db"}
+{"spark": "331"}
+{"spark": "332"}
+{"spark": "333"}
+spark-rapids-shim-json-lines ***/
+package org.apache.spark.sql.hive.rapids.shims
+
+// Base trait from which all hive insert statement physical execution extends.
+private[hive] trait GpuSaveAsHiveFile extends GpuDataWritingCommand with SaveAsHiveFile {
+
+  // TODO(future): Examine compressions options.
+  // - Apache Spark 3.1-3 has code to examine Hadoop compression settings
+  //   (and takes a FileSinkDesc instead of FileFormat).
+  // - Apache Spark 3.4 has removed all that logic.
+  // - GPU Hive text writer does not support compression for output.
+  protected def gpuSaveAsHiveFile(sparkSession: SparkSession,
+                                  plan: SparkPlan,
+                                  hadoopConf: Configuration,
+                                  fileFormat: ColumnarFileFormat,
+                                  outputLocation: String,
+                                  customPartitionLocations: Map[TablePartitionSpec,String] = Map.empty,
+                                  partitionAttributes: Seq[Attribute] = Nil,
+                                  bucketSpec: Option[BucketSpec] = None,
+                                  options: Map[String, String] = Map.empty): Set[String] = {
+
+    val committer = FileCommitProtocol.instantiate(
+      sparkSession.sessionState.conf.fileCommitProtocolClass,
+      jobId = java.util.UUID.randomUUID().toString,
+      outputPath = outputLocation)
+
+    GpuFileFormatWriter.write(
+      sparkSession = sparkSession,
+      plan = plan,
+      fileFormat = fileFormat,
+      committer = committer,
+      outputSpec =
+        FileFormatWriter.OutputSpec(outputLocation, customPartitionLocations, outputColumns),
+      hadoopConf = hadoopConf,
+      partitionColumns = partitionAttributes,
+      bucketSpec = bucketSpec,
+      statsTrackers = Seq(gpuWriteJobStatsTracker(hadoopConf)),
+      options = options,
+      useStableSort = false,                  // TODO: Fetch from RapidsConf.
+      concurrentWriterPartitionFlushSize = 0L // TODO: Fetch from RapidsConf.
+    )
+  }
+}
